requirements.txt
google-generativeai
python-dotenv
peft
trl
transformers
datasets
bitsandbytes
accelerate
llama-cpp-python
gradio
